Separate QAveragePooling2D
